+++ .pipelines/pipeline.user.windows.yml
@@ -72,6 +72,7 @@ build:
             - 'modules\ColorPicker\ColorPickerUI.dll'
             - 'modules\ColorPicker\ColorPickerUI.exe'
             - 'modules\ColorPicker\ManagedCommon.dll'
+            - 'modules\ColorPicker\Microsoft.PowerToys.Common.UI.dll'
             - 'modules\ColorPicker\Microsoft.PowerToys.Settings.UI.Lib.dll'
             - 'modules\ColorPicker\PowerToysInterop.dll'
             - 'modules\ColorPicker\Telemetry.dll'
@@ -94,9 +95,13 @@ build:
             - 'modules\ImageResizer\ImageResizer.dll'
             - 'modules\ImageResizer\ImageResizerExt.dll'
             - 'modules\ImageResizer\PowerToysInterop.dll'
+            - 'modules\ImageResizer\ManagedCommon.dll'
+            - 'modules\ImageResizer\ManagedTelemetry.dll'
+            - 'modules\ImageResizer\Microsoft.PowerToys.Common.UI.dll'
             - 'modules\KeyboardManager\KeyboardManager.dll'
             - 'modules\launcher\Microsoft.PowerToys.Settings.UI.Lib.dll'
             - 'modules\launcher\ManagedCommon.dll'
+            - 'modules\launcher\Microsoft.PowerToys.Common.UI.dll'
             - 'modules\launcher\Microsoft.Launcher.dll'
             - 'modules\launcher\Plugins\Microsoft.Plugin.Calculator\Microsoft.Plugin.Calculator.dll'
             - 'modules\launcher\Plugins\Microsoft.Plugin.Calculator\Wox.Infrastructure.dll'
