+++ src/common/Microsoft.PowerToys.Common.UI/ThemeManager.cs
@@ -1,14 +1,14 @@
-﻿// Copyright (c) Microsoft Corporation
+// Copyright (c) Microsoft Corporation
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
 using System;
 using System.Linq;
 using System.Windows;
-using ControlzEx.Theming;
+using ManagedCommon;
 using Microsoft.Win32;
 
-namespace ImageResizer.Utilities
+namespace Microsoft.PowerToys.Common.UI
 {
     public class ThemeManager : IDisposable
     {
@@ -37,32 +37,32 @@ namespace ImageResizer.Utilities
             Uri darkThemeUri = new Uri("pack://application:,,,/Themes/Dark.xaml");
 
             ControlzEx.Theming.ThemeManager.Current.AddLibraryTheme(
-                new LibraryTheme(
+                new ControlzEx.Theming.LibraryTheme(
                     highContrastOneThemeUri,
                     CustomLibraryThemeProvider.DefaultInstance));
             ControlzEx.Theming.ThemeManager.Current.AddLibraryTheme(
-                new LibraryTheme(
+                new ControlzEx.Theming.LibraryTheme(
                     highContrastTwoThemeUri,
                     CustomLibraryThemeProvider.DefaultInstance));
             ControlzEx.Theming.ThemeManager.Current.AddLibraryTheme(
-                new LibraryTheme(
+                new ControlzEx.Theming.LibraryTheme(
                     highContrastBlackThemeUri,
                     CustomLibraryThemeProvider.DefaultInstance));
             ControlzEx.Theming.ThemeManager.Current.AddLibraryTheme(
-                new LibraryTheme(
+                new ControlzEx.Theming.LibraryTheme(
                     highContrastWhiteThemeUri,
                     CustomLibraryThemeProvider.DefaultInstance));
             ControlzEx.Theming.ThemeManager.Current.AddLibraryTheme(
-                new LibraryTheme(
+                new ControlzEx.Theming.LibraryTheme(
                     lightThemeUri,
                     CustomLibraryThemeProvider.DefaultInstance));
             ControlzEx.Theming.ThemeManager.Current.AddLibraryTheme(
-                new LibraryTheme(
+                new ControlzEx.Theming.LibraryTheme(
                     darkThemeUri,
                     CustomLibraryThemeProvider.DefaultInstance));
 
             ResetTheme();
-            ControlzEx.Theming.ThemeManager.Current.ThemeSyncMode = ThemeSyncMode.SyncWithAppMode;
+            ControlzEx.Theming.ThemeManager.Current.ThemeSyncMode = ControlzEx.Theming.ThemeSyncMode.SyncWithAppMode;
             ControlzEx.Theming.ThemeManager.Current.ThemeChanged += Current_ThemeChanged;
             SystemParameters.StaticPropertyChanged += SystemParameters_StaticPropertyChanged;
         }
@@ -97,73 +97,83 @@ namespace ImageResizer.Utilities
                 case "hcblack":
                     return Theme.HighContrastBlack;
                 default:
-                    return Theme.None;
+                    return Theme.HighContrastOne;
             }
         }
 
         private void ResetTheme()
         {
-            if (SystemParameters.HighContrast)
-            {
-                Theme highContrastBaseType = GetHighContrastBaseType();
-                ChangeTheme(highContrastBaseType);
-            }
-            else
-            {
-                string baseColor = WindowsThemeHelper.GetWindowsBaseColor();
-                ChangeTheme((Theme)Enum.Parse(typeof(Theme), baseColor));
-            }
+            ChangeTheme(currentTheme, false);
         }
 
-        private void ChangeTheme(Theme theme)
+        public void ChangeTheme(Theme theme, bool forceSystem)
         {
             Theme oldTheme = currentTheme;
-            if (theme == currentTheme)
-            {
-                return;
-            }
 
-            if (theme == Theme.HighContrastOne)
+            if (theme == Theme.System)
             {
-                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastOneTheme);
+                currentTheme = Theme.System;
+                if (ControlzEx.Theming.WindowsThemeHelper.IsHighContrastEnabled())
+                {
+                    Theme highContrastBaseType = GetHighContrastBaseType();
+                    ChangeTheme(highContrastBaseType, true);
+                }
+                else
+                {
+                    string baseColor = ControlzEx.Theming.WindowsThemeHelper.GetWindowsBaseColor();
+                    ChangeTheme((Theme)Enum.Parse(typeof(Theme), baseColor), true);
+                }
+            }
+            else if (theme == Theme.HighContrastOne)
+            {
                 currentTheme = Theme.HighContrastOne;
+                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastOneTheme, true);
             }
             else if (theme == Theme.HighContrastTwo)
             {
-                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastTwoTheme);
                 currentTheme = Theme.HighContrastTwo;
+                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastTwoTheme, true);
             }
             else if (theme == Theme.HighContrastWhite)
             {
-                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastWhiteTheme);
                 currentTheme = Theme.HighContrastWhite;
+                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastWhiteTheme, true);
             }
             else if (theme == Theme.HighContrastBlack)
             {
-                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastBlackTheme);
                 currentTheme = Theme.HighContrastBlack;
+                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, HighContrastBlackTheme, true);
             }
             else if (theme == Theme.Light)
             {
-                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, LightTheme);
                 currentTheme = Theme.Light;
+                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, LightTheme);
             }
             else if (theme == Theme.Dark)
             {
-                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, DarkTheme);
                 currentTheme = Theme.Dark;
-            }
-            else
-            {
-                currentTheme = Theme.None;
+                ControlzEx.Theming.ThemeManager.Current.ChangeTheme(_app, DarkTheme);
             }
 
             ThemeChanged?.Invoke(oldTheme, currentTheme);
+
+            if (forceSystem)
+            {
+                currentTheme = Theme.System;
+            }
         }
 
-        private void Current_ThemeChanged(object sender, ThemeChangedEventArgs e)
+        private void Current_ThemeChanged(object sender, ControlzEx.Theming.ThemeChangedEventArgs e)
         {
-            ResetTheme();
+            ControlzEx.Theming.ThemeManager.Current.ThemeChanged -= Current_ThemeChanged;
+            try
+            {
+                ResetTheme();
+            }
+            finally
+            {
+                ControlzEx.Theming.ThemeManager.Current.ThemeChanged += Current_ThemeChanged;
+            }
         }
 
         protected virtual void Dispose(bool disposing)
@@ -173,7 +183,6 @@ namespace ImageResizer.Utilities
                 if (disposing)
                 {
                     ControlzEx.Theming.ThemeManager.Current.ThemeChanged -= Current_ThemeChanged;
-                    SystemParameters.StaticPropertyChanged -= SystemParameters_StaticPropertyChanged;
                     _disposed = true;
                 }
             }
@@ -187,15 +196,4 @@ namespace ImageResizer.Utilities
     }
 
     public delegate void ThemeChangedHandler(Theme oldTheme, Theme newTheme);
-
-    public enum Theme
-    {
-        None,
-        Light,
-        Dark,
-        HighContrastOne,
-        HighContrastTwo,
-        HighContrastBlack,
-        HighContrastWhite,
-    }
 }
