+++ src/modules/launcher/PowerLauncher/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application x:Class="PowerLauncher.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:theming="clr-namespace:Wox.Plugin;assembly=Wox.Plugin"
+             xmlns:theming="clr-namespace:Microsoft.PowerToys.Common.UI;assembly=Microsoft.PowerToys.Common.UI"
              ShutdownMode="OnMainWindowClose"
              Startup="OnStartup">
     <Application.Resources>
