+++ src/modules/launcher/PowerLauncher/App.xaml.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Windows;
 using ManagedCommon;
 using Microsoft.PowerLauncher.Telemetry;
+using Microsoft.PowerToys.Common.UI;
 using Microsoft.PowerToys.Telemetry;
 using PowerLauncher.Helper;
 using PowerLauncher.Plugin;
