+++ src/modules/launcher/PowerLauncher/PublicAPIInstance.cs
@@ -8,10 +8,10 @@ using System.Linq;
 using System.Net;
 using System.Windows;
 using ManagedCommon;
+using Microsoft.PowerToys.Common.UI;
 using PowerLauncher.Helper;
 using PowerLauncher.Plugin;
 using PowerLauncher.ViewModel;
-using Wox.Infrastructure;
 using Wox.Infrastructure.Image;
 using Wox.Plugin;
