+++ src/modules/launcher/Wox.Plugin/IPublicAPI.cs
@@ -4,6 +4,7 @@
 
 using System.Collections.Generic;
 using ManagedCommon;
+using Microsoft.PowerToys.Common.UI;
 
 namespace Wox.Plugin
 {
